requirements.txt
pylint~=2.15.5
pylint-pydantic~=0.1.4
click~=8.1.3
numpy~=1.23.3
mlflow~=1.30.0
nibabel~=4.0.2
p_tqdm~=1.4.0
pandas~=1.5.0
pytest~=7.2.0
python-dotenv~=0.21.0
scikit_learn~=1.1.2
scipy~=1.9.3
setuptools~=65.5.0
matplotlib~=3.6.2
seaborn~=0.12.1
torch~=1.13.0
torchvision~=0.14.0
captum~=0.5.0
torchmetrics~=0.10.0
pytorch_lightning~=1.7.7
kornia~=0.6.8
jupyter~=1.0.0
fastapi~=0.86.0

starlette~=0.20.4
pydantic~=1.10.2
scikit-learn~=1.1.3
celery~=5.2.7
uvicorn~=0.19.0
redis~=4.3.4
python-multipart~=0.0.5
nibabel~=4.0.2

pyyaml~=6.0
ua-parser~=0.16.1
user-agents~=2.2.0

locust~=2.14.0
